+++ packages/vue/src/components/editable/editable-input.vue
@@ -20,5 +20,5 @@ const editable = useEditableContext()
 </script>
 
 <template>
-  <ark.input v-bind="editable.getInputProps()" :as-child="asChild" />
+  <ark.input v-bind="editable.getInputProps()" :as-child="asChild"><slot /></ark.input>
 </template>