+++ packages/vue/src/components/file-upload/file-upload-hidden-input.vue
@@ -20,5 +20,5 @@ const fileUpload = useFileUploadContext()
 </script>
 
 <template>
-  <ark.input v-bind="fileUpload.getHiddenInputProps()" :as-child="asChild" />
+  <ark.input v-bind="fileUpload.getHiddenInputProps()" :as-child="asChild"><slot /></ark.input>
 </template>