+++ packages/vue/src/components/number-input/number-input-input.vue
@@ -20,5 +20,5 @@ const numberInput = useNumberInputContext()
 </script>
 
 <template>
-  <ark.input v-bind="numberInput.getInputProps()" :as-child="asChild" />
+  <ark.input v-bind="numberInput.getInputProps()" :as-child="asChild"><slot /></ark.input>
 </template>