+++ packages/vue/src/components/pin-input/pin-input-hidden-input.vue
@@ -20,5 +20,5 @@ const pinInput = usePinInputContext()
 </script>
 
 <template>
-  <ark.input v-bind="pinInput.getHiddenInputProps()" :as-child="asChild" />
+  <ark.input v-bind="pinInput.getHiddenInputProps()" :as-child="asChild"><slot /></ark.input>
 </template>