+++ packages/vue/src/components/tags-input/tags-input-hidden-input.vue
@@ -20,5 +20,5 @@ const tagsInput = useTagsInputContext()
 </script>
 
 <template>
-  <ark.input v-bind="tagsInput.getHiddenInputProps()" :as-child="asChild" />
+  <ark.input v-bind="tagsInput.getHiddenInputProps()" :as-child="asChild"><slot /></ark.input>
 </template>